+++ src/doomstat.h
@@ -497,6 +497,9 @@ extern int wipe_speed_percentage;
 extern int alt_interpic;
 extern int s_clipping_dist_x2;
 extern int one_key_saveload;
+extern int rewind_interval;
+extern int rewind_depth;
+extern int rewind_timeout;
 extern int no_page_ticking;
 extern int quick_quitgame;