+++ src/m_menu.c
@@ -3157,6 +3157,8 @@ enum {
   keys10_xhair,
   keys10_stub4,
   keys10_lastweap,
+  keys10_stub5,
+  keys10_rewind,
 };
 
 setup_menu_t keys_settings10[] =
@@ -3174,6 +3176,8 @@ setup_menu_t keys_settings10[] =
     {"Toggle Crosshair", S_INPUT,                     m_scrn, KB_X, M_Y + keys10_xhair    * M_SPC, {0}, input_crosshair},
     {"",                 S_SKIP,                      m_null, KB_X, M_Y + keys10_stub4    * M_SPC},
     {"Last Used Weapon", S_INPUT,                     m_scrn, KB_X ,M_Y + keys10_lastweap * M_SPC, {0}, input_lastweapon},
+    {"",                 S_SKIP,                      m_null, KB_X, M_Y + keys10_stub5    * M_SPC},
+    {"Rewind",           S_INPUT|S_STRICT|S_CRITICAL, m_scrn, KB_X ,M_Y + keys10_rewind   * M_SPC, {0}, input_rewind},
 
   {"<- PREV", S_SKIP|S_PREV, m_null, M_X_PREV, M_Y_PREVNEXT, {keys_settings9}},
   {"NEXT ->", S_SKIP|S_NEXT, m_null, M_X_NEXT, M_Y_PREVNEXT, {keys_settings11}},
@@ -4603,6 +4607,9 @@ enum {
   gen8_sclipdist,
   gen8_organizesaves,
   gen8_quicksaveload,
+  gen8_rwinterval,
+  gen8_rwdepth,
+  gen8_rwtimeout,
   gen8_nopagetic,
   gen8_quickexit,
   gen8_stub1,
@@ -4960,6 +4967,18 @@ setup_menu_t gen_settings7[] = {
   {0,S_SKIP|S_END,m_null}
 };
 
+static void M_UpdateRewindInterval(void)
+{
+  G_EnableRewind();
+  G_ResetRewindCountdown();
+}
+
+static void M_UpdateRewindDepth(void)
+{
+  G_EnableRewind();
+  G_ClearExcessKeyFrames();
+}
+
 static const char *s_clipping_dists[] = {
   "1200", "2400", NULL
 };
@@ -4972,11 +4991,14 @@ setup_menu_t gen_settings8[] = { // [Nugget]
 
   {"Nugget - Miscellaneous", S_SKIP|S_TITLE, m_null, M_X, M_Y + gen8_title1 * M_SPC},
 
-    {"Sound Hearing Distance",  S_CHOICE|S_STRICT, m_null, M_X, M_Y + gen8_sclipdist     * M_SPC, {"s_clipping_dist_x2"}, 0, M_SetSoundModule, s_clipping_dists},
-    {"Organize Saves by IWAD",  S_YESNO|S_PRGWARN, m_null, M_X, M_Y + gen8_organizesaves * M_SPC, {"organize_saves"}},
-    {"One-Key Quick Save/Load", S_YESNO,           m_null, M_X, M_Y + gen8_quicksaveload * M_SPC, {"one_key_saveload"}},
-    {"Play Internal Demos",     S_CHOICE,          m_null, M_X, M_Y + gen8_nopagetic     * M_SPC, {"no_page_ticking"}, 0, NULL, page_ticking_conds},
-    {"Quick \"Quit Game\"",     S_YESNO,           m_null, M_X, M_Y + gen8_quickexit     * M_SPC, {"quick_quitgame"}},
+    {"Sound Hearing Distance",  S_CHOICE|S_STRICT,         m_null, M_X, M_Y + gen8_sclipdist     * M_SPC, {"s_clipping_dist_x2"}, 0, M_SetSoundModule, s_clipping_dists},
+    {"Organize Saves by IWAD",  S_YESNO |S_PRGWARN,        m_null, M_X, M_Y + gen8_organizesaves * M_SPC, {"organize_saves"}},
+    {"One-Key Quick Save/Load", S_YESNO,                   m_null, M_X, M_Y + gen8_quicksaveload * M_SPC, {"one_key_saveload"}},
+    {"Rewind Interval (S)",     S_NUM|S_STRICT|S_CRITICAL, m_null, M_X, M_Y + gen8_rwinterval    * M_SPC, {"rewind_interval"}, 0, M_UpdateRewindInterval},
+    {"Rewind Depth",            S_NUM|S_STRICT|S_CRITICAL, m_null, M_X, M_Y + gen8_rwdepth       * M_SPC, {"rewind_depth"}, 0, M_UpdateRewindDepth},
+    {"Rewind Timeout (MS)",     S_NUM|S_STRICT|S_CRITICAL, m_null, M_X, M_Y + gen8_rwtimeout     * M_SPC, {"rewind_timeout"}, 0, G_EnableRewind},
+    {"Play Internal Demos",     S_CHOICE,                  m_null, M_X, M_Y + gen8_nopagetic     * M_SPC, {"no_page_ticking"}, 0, NULL, page_ticking_conds},
+    {"Quick \"Quit Game\"",     S_YESNO,                   m_null, M_X, M_Y + gen8_quickexit     * M_SPC, {"quick_quitgame"}},
 
   {"",                       S_SKIP,         m_null, M_X, M_Y + gen8_stub1  * M_SPC},
   {"Nugget - Accessibility", S_SKIP|S_TITLE, m_null, M_X, M_Y + gen8_title2 * M_SPC},
@@ -6322,7 +6344,8 @@ boolean M_Responder (event_t* ev)
     {                                                         //  |
       static boolean fastdemo_timer = false;                  //  V
 
-      // [Nugget]
+      // [Nugget] /-----------------------------------------------------------
+
       if (M_InputActivated(input_crosshair))
 	{
 	  extern void HU_StartCrosshair(void);
@@ -6332,11 +6355,18 @@ boolean M_Responder (event_t* ev)
 	  togglemsg("Crosshair %s", hud_crosshair_on ? "Enabled" : "Disabled");
 	}
 
-      // [Nugget]
       if (STRICTMODE(M_InputActivated(input_chasecam)))
 	{
 	  if (++chasecam_mode > CHASECAMMODE_FRONT) { chasecam_mode = CHASECAMMODE_OFF; }
 	}
+
+      if (STRICTMODE(M_InputActivated(input_rewind)))
+	{
+	  G_Rewind();
+	  return true;
+	}
+
+      // [Nugget] -----------------------------------------------------------/
 
       if (M_InputActivated(input_autorun)) // Autorun         //  V
 	{